+++ src/mcu_digital.h
@@ -81,7 +81,7 @@ void GPIO_SetPinDirection(GPIO_TypeDef *gpio, uint32_t gpioNum, GPIO_PinDirectio
  * \param mask Маска выводов
  * \param dir Направление выводов
  */
-void GPIO_SetPinMaskDirection(GPIO_TypeDef *gpio, uint32_t mask, GPIO_PinDirection dir);
+void GPIO_SetPinMaskDirection(GPIO_TypeDef *gpio, uint16_t mask, GPIO_PinDirection dir);
 
 
 /**
